refactor(wasm): move il2cpp line mapping into its own module
Address review: mirror the Rust crate layout instead of hanging an
Il2cpp-specific operation off the generic ObjectFile. The binding now
lives in a dedicated `il2cpp` module as a free function
`il2cppLineMapping(object, provider)`, paralleling the standalone
`symbolic_il2cpp::ObjectLineMapping`.

Add a `pub(crate)` accessor on Object so sibling modules can borrow the
underlying debuginfo object. Move the wasm-bindgen test into a matching
tests/il2cpp.rs.
